+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Change Log
 ==========
+                
+## Next version
+
+* Note: TMDB has added additional primary translation languages: `ku-TR`, `so-SO` and `uz-UZ`. 
 
 ## 2.11.0
 _2024-06-07_